(almost) fix compilation under -Wall (gcc and clang) for cpp/hdf; two warnings remain
- removal of unused variables
- int -> size_t changes
- Fix incorrect signature for overriding StoreHoleStatus

Major remaining warning involves confusing interaction of overloading and overriding of HDFData::Initialize.

Comments containing "FIXME(yli)" should be reviewed by yli to see if anything else needs to be changed.

[Bug 26302]

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/lib/cpp/": change = 147809]

+++ hdf/HDFUtils.cpp
@@ -7,7 +7,7 @@ std::string GetH5MovieName(std::string fileName) {
 
 std::vector<std::string> GetH5MovieNames(const std::vector<std::string> & fileNames) {
     std::vector<std::string> ret;
-    for (int i = 0 ; i < fileNames.size(); i++) {
+    for (size_t i = 0 ; i < fileNames.size(); i++) {
         ret.push_back(GetH5MovieName(fileNames[i]));
     }
     return ret;
@@ -16,7 +16,7 @@ std::vector<std::string> GetH5MovieNames(const std::vector<std::string> & fileNa
 std::vector< std::pair<UInt, UInt> > GetMinMaxHoleNumbers(
     const std::vector<std::string> & fileNames, bool isRGN) {
     std::vector< std::pair<UInt, UInt> > ret;
-    for (int i = 0 ; i < fileNames.size(); i++) {
+    for (size_t i = 0 ; i < fileNames.size(); i++) {
         ret.push_back(GetMinMaxHoleNumber(fileNames[i], isRGN));
     }
     return ret;
@@ -60,8 +60,8 @@ std::vector<int> MapPls2Rgn(const std::vector<std::string> & plsFNs,
     std::vector< std::pair<UInt, UInt> > rgnHoles = GetMinMaxHoleNumbers(rgnFNs, true);
 
     std::vector<int> ret;
-    for (int i = 0; i < plsFNs.size(); i++) {
-        int j = 0;
+    for (size_t i = 0; i < plsFNs.size(); i++) {
+        size_t j = 0;
         for (; j < rgnFNs.size(); j++) {
             if (plsMovies[i] == rgnMovies[j] and
                 plsHoles[i].first <= rgnHoles[j].first and
